requirements.txt
numpy==1.24.2
opencv_python==4.7.0.72
pandas==1.5.3
tensorflow==2.11.0
tensorflow_intel==2.11.0

pillow~=9.4.0
opencv-python~=4.7.0.72
keras~=2.11.0
matplotlib~=3.7.1
utils~=1.0.1
scikit-learn~=1.2.2
albumentations~=1.3.1